requirements.txt
gnureadline==6.3.3
ipython==2.4.0
matplotlib==1.4.2
nose==1.3.4
numpy==1.9.1
pyparsing==2.0.3
pyserial==2.7
PySide==1.2.2
python-dateutil==2.4.0
pytz==2014.10
six==1.9.0
SQLAlchemy==0.9.8
